minirt final version 100

+++ minirt/srcs/math/math_1.c
@@ -1,18 +1,18 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   math.c                                             :+:      :+:    :+:   */
+/*   math_1.c                                           :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: ekadiri <ekadiri@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/06/08 12:26:41 by ekadiri           #+#    #+#             */
-/*   Updated: 2023/06/29 22:02:34 by ekadiri          ###   ########.fr       */
+/*   Updated: 2023/09/22 16:47:57 by ekadiri          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../minirt.h"
 
-float	distance(t_coord v1, t_coord v2)
+double	distance(t_coord v1, t_coord v2)
 {
 	return (sqrt(pow(v1.x - v2.x, 2.) + pow(v1.y - v2.y, 2.)
 			+ pow(v1.z - v2.z, 2.)));
@@ -20,7 +20,7 @@ float	distance(t_coord v1, t_coord v2)
 
 t_coord	norm(t_coord v)
 {
-	float	dst;
+	double	dst;
 
 	dst = distance(coord(0., 0., 0.), v);
 	return (coord(v.x / dst, v.y / dst, v.z / dst));
